requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-core==2.2.3
google-cloud-datastore==2.5.1
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
